# openapi.reference.yml
openapi: 3.0.2
info:
  title: DenBar
  version: '1.0'
  description: Project to order drinks to be delivered
  contact:
    name: Grasshopper.io
    email: info@grasshopper.io
    url: 'https://grasshopper.io'
servers:
  - url: '{server}/api'
    variables:
      server:
        enum:
          - 'http://localhost:5000'
        description: 'When running locally, select localhost'
        default: 'http://localhost:5000'
paths:
...
  '/products/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: getProductById
      tags:
        - Post
      responses: {}
      operationId: get-posts-id
    delete:
      summary: deletePost
      tags:
        - Post
      responses: {}
      operationId: delete-posts-id
    patch:
      summary: modifyPost
      tags:
        - Post
      responses: {}
      operationId: patch-posts-id
...
components:
  securitySchemes:
    Basic authentication:
      type: http
      scheme: basic
      description: Basic authentication for user actions
...
security:
  - Basic authentication: []